@feature user, patiets get have implemented

+++ middlewares/jwt-validation.middleware.js
@@ -7,7 +7,7 @@ const isJwtValid = async (req, resp = response, next) => {
     const token = req.header('x-token');
     const file = req.file
     if (!token) { 
-        await fs.unlink(file.path)
+        if (file) { await fs.unlink(file.path) }
         return resp.status(401).json({
             ok: false,
             message: ' No token was sent by user'
@@ -18,7 +18,7 @@ const isJwtValid = async (req, resp = response, next) => {
         req.user_id = user_id;
         next();
     } catch (error) {
-        await fs.unlink(file.path)
+        if (file) { await fs.unlink(file.path) }
         return resp.status(403).json({
             ok: false,
             message:'Invalid token'
